fix bad imports from src/core/port.ts
