+++ src/command/preview/preview.ts
@@ -67,13 +67,14 @@ import {
 import {
   kProjectWatchInputs,
   ProjectContext,
-  resolvePreviewOptions,
+  ProjectPreview,
 } from "../../project/types.ts";
 import { projectOutputDir } from "../../project/project-shared.ts";
 import { projectContext } from "../../project/project-context.ts";
 import { normalizePath, pathWithForwardSlashes } from "../../core/path.ts";
 import {
   isJupyterHubServer,
+  isRStudio,
   isRStudioServer,
   isRStudioWorkbench,
   isVSCodeServer,
@@ -101,6 +102,38 @@ import {
   kPreviewModeRaw,
 } from "../../config/constants.ts";
 import { isJatsOutput } from "../../config/format.ts";
+import { mergeConfigs } from "../../core/config.ts";
+import { kLocalhost } from "../../core/port-consts.ts";
+import { findOpenPort, waitForPort } from "../../core/port.ts";
+
+export async function resolvePreviewOptions(
+  options: ProjectPreview,
+  project?: ProjectContext,
+): Promise<ProjectPreview> {
+  // start with project options if we have them
+  if (project?.config?.project.preview) {
+    options = mergeConfigs(project.config.project.preview, options);
+  }
+  // provide defaults
+  const resolved = mergeConfigs({
+    host: kLocalhost,
+    browser: true,
+    [kProjectWatchInputs]: !isRStudio(),
+    timeout: 0,
+    navigate: true,
+  }, options) as ProjectPreview;
+
+  // if a specific port is requested then wait for it up to 5 seconds
+  if (resolved.port) {
+    if (!await waitForPort({ port: resolved.port, hostname: resolved.host })) {
+      throw new Error(`Requested port ${options.port} is already in use.`);
+    }
+  } else {
+    resolved.port = findOpenPort();
+  }
+
+  return resolved;
+}
 
 interface PreviewOptions {
   port?: number;
