+++ src/core/http-devserver.ts
@@ -12,7 +12,7 @@ import { renderEjs } from "./ejs.ts";
 import { maybeDisplaySocketError } from "./http.ts";
 import { FileResponse } from "./http-types.ts";
 import { LogEventsHandler } from "./log.ts";
-import { kLocalhost } from "./port.ts";
+import { kLocalhost } from "./port-consts.ts";
 import { resourcePath } from "./resources.ts";
 import { isRStudioPreview, isRStudioServer } from "./platform.ts";
 import { kTextHtml } from "./mime.ts";
