+++ src/core/port-consts.ts
@@ -1,0 +1,7 @@
+/*
+ * port-consts.ts
+ *
+ * Copyright (C) 2020-2023 Posit Software, PBC
+ */
+
+export const kLocalhost = "127.0.0.1";
