+++ src/core/port.ts
@@ -1,16 +1,15 @@
 /*
-* port.ts
-*
-* Copyright (C) 2020-2022 Posit Software, PBC
-*
-*/
+ * port.ts
+ *
+ * Copyright (C) 2020-2022 Posit Software, PBC
+ */
 
 import * as ld from "./lodash.ts";
 
 import { randomInt } from "./random.ts";
 import { sleep } from "./wait.ts";
 
-export const kLocalhost = "127.0.0.1";
+import { kLocalhost } from "./port-consts.ts";
 
 const kMinPort = 3000;
 const kMaxPort = 8000;
