+++ src/project/serve/serve.ts
@@ -24,8 +24,9 @@ import {
   kProject404File,
   kProjectType,
   ProjectContext,
-  resolvePreviewOptions,
 } from "../../project/types.ts";
+import { resolvePreviewOptions } from "../../command/preview/preview.ts";
+
 import {
   isProjectInputFile,
   projectExcludeDirs,
@@ -104,7 +105,8 @@ import { projectScratchPath } from "../project-scratch.ts";
 import { monitorPreviewTerminationConditions } from "../../core/quarto.ts";
 import { exitWithCleanup, onCleanup } from "../../core/cleanup.ts";
 import { projectExtensionDirs } from "../../extension/extension.ts";
-import { findOpenPort, kLocalhost } from "../../core/port.ts";
+import { findOpenPort } from "../../core/port.ts";
+import { kLocalhost } from "../../core/port-consts.ts";
 import { ProjectServe } from "../../resources/types/schema-types.ts";
 import { handleHttpRequests } from "../../core/http-server.ts";
 import { touch } from "../../core/file.ts";
